src/requirements.txt
django-cors-headers==3.10.0
django-environ==0.7.0
djangorestframework==3.12.4
psycopg2-binary==2.9.1

# for tests
mypy==0.910
pytest==6.2.5
pytest-cov==3.0.0
pytest-django==4.4.0
pytest-flake8==1.0.7
responses==0.14.0
